audio: Use a common base for audio firmware
