+++ audio/in/descriptors.strings
