+++ common/common.mk
@@ -83,7 +83,8 @@ check: check_int2jt
 clean:
 	$(Q_RM)$(RM) *.adb *.asm *.cdb *.iic *.lk *.lnk *.lst *.omf *.map \
 		*.mem *.rel *.rst *.sym descriptors_strings.* a.out date.h \
-		date.inc progOffsets.h version_data.h version_data.c ${TARGET}.hex
+		date.inc progOffsets.h ${TARGET}.hex $(FIRMWARE)/*.asm \
+		$(FIRMWARE)/*.lst $(FIRMWARE)/*.rel $(FIRMWARE)/*.rst $(FIRMWARE)/*.sym
 	cd $(FX2LIBDIR) && make clean
 
 distclean: clean
